+++ client/src/main/kotlin/settings/Setting.kt
@@ -1,3 +1,3 @@
-package preference
+package settings
 
 data class Setting<T>(val name: String, val default: T)
