+++ client/src/main/kotlin/settings/SettingChangeListener.kt
@@ -1,0 +1,5 @@
+package settings
+
+interface SettingChangeListener {
+    fun <T> settingChanged(setting: Setting<T>, newValue: T?)
+}
